change api docs

+++ controllers/passController.js
@@ -42,8 +42,12 @@ class PassController{
         Pass
             .findByPk(req.params.id)
             .then(data1 => {
-                        results = Object.assign(data1)
-                        return Pass.destroy({where : {id : req.params.id},returning : true})
+                        if(data1){
+                            results = Object.assign(data1)
+                            return Pass.destroy({where : {id : req.params.id},returning : true})
+                        }else{
+                            res.status(404).json({error : 'not found'})
+                        }                  
                         })
             .then(data2 => {
                 res.status(200).json({todo : results})
